[Modules] Update modules & utils (#252)
