+++ cogdl/models/__init__.py
@@ -75,8 +75,8 @@ SUPPORTED_MODELS = {
     "line": "cogdl.models.emb.line",
     "sdne": "cogdl.models.emb.sdne",
     "prone": "cogdl.models.emb.prone",
-    "daegc": "cogdl.models.agc.daegc",
-    "agc": "cogdl.models.agc.agc",
+    "daegc": "cogdl.models.nn.daegc",
+    "agc": "cogdl.models.nn.agc",
     "gae": "cogdl.models.nn.gae",
     "vgae": "cogdl.models.nn.gae",
     "dgi": "cogdl.models.nn.dgi",
