+++ cogdl/utils/__init__.py
@@ -1,3 +1,5 @@
 from .utils import *
 from .evaluator import *
 from .sampling import *
+from .graph_utils import *
+from .spmm_utils import *
