+++ cogdl/utils/link_prediction_utils.py
