+++ cogdl/utils/prone_utils.py
@@ -2,7 +2,6 @@ import math
 import numpy as np
 import scipy
 import scipy.sparse as sp
-from scipy.sparse.linalg import expm
 from scipy.special import iv
 from sklearn import preprocessing
